pyproject.toml
[project]
name = "lesson26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "intersystems-irispython>=5.1.2",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
